Duke3d: reduce use of temporary variables in a few VM opcodes
Also moves a bunch of VM_ASSERT statements up next to the variables they check.

When this macro was first introduced, all of an instruction's operands needed to be read from the bytecode before aborting due to failure. Since this isn't the case anymore, all of the VM_ASSERTs can be moved up.